stresschecker/requirements.txt
flask
flask-sqlalchemy
gunicorn
werkzeug
qrcode[pil]
fpdf2
